Organize imports. Extracted create about to ObservationManagerMenuHelp
